fix: update forum post and comment

+++ app/forum.py
@@ -12,12 +12,22 @@ def create_post():
     data = request.get_json()
     user_id = get_jwt_identity()
     post_text = data.get('post_text')
-    post_image = str(data.get('post_image'))
+    post_image = data.get('post_image')
 
     try:
+        user_ref = db.collection('users').document(user_id)
+        user_doc = user_ref.get()
+        if not user_doc.exists:
+            raise Exception("User not found")
+        
+        user_data = user_doc.to_dict()
+        
         doc_ref = db.collection('forum').document()
         doc_ref.set({
             "user_id": user_id,
+            "user_image": user_data.get('imageProfile', ''),
+            "user_name": user_data.get('name', ''),
+            "user_email": user_data.get('email', ''),
             "post_text": post_text,
             "post_image": post_image,
             "count_like": 0,
@@ -71,8 +81,18 @@ def get_post(post_id):
 @forum_bp.route('/forum/<post_id>', methods=['DELETE'])
 @jwt_required()
 def delete_post(post_id):
+    user_id = get_jwt_identity()
     try:
         post_ref = db.collection('forum').document(post_id)
+        post_doc = post_ref.get()
+
+        if not post_doc.exists:
+            raise Exception("Post not found")
+
+        post_data = post_doc.to_dict()
+        if post_data["user_id"] != user_id:
+            return jsonify({"message": "You are not authorized to delete this post"}), 403
+
         post_ref.delete()
         return jsonify({"message": "post has been deleted"}), 200
     except Exception as e:
@@ -129,9 +149,12 @@ def create_comment():
             "created_at": int(datetime.now().timestamp())
         })
 
-        post_data = post_doc.to_dict()
         comments = post_ref.collection('comments').stream()
-        comments_list = [{"user_id": comment.get('user_id'), "comment_id": comment.id, "comment": comment.get('comment'), "created_at": comment.get('created_at')} for comment in comments]
+        comments_list = [{"user_id": comment.get('user_id'), 
+                          "comment_id": comment.id, 
+                          "comment": comment.get('comment'), 
+                          "created_at": comment.get('created_at')} 
+                          for comment in comments]
 
         return jsonify({"post_id": post_id, "comments": comments_list}), 201
     except Exception as e:
@@ -141,8 +164,24 @@ def create_comment():
 @forum_bp.route('/comment/<comment_id>', methods=['DELETE'])
 @jwt_required()
 def delete_comment(comment_id):
+    user_id = get_jwt_identity()
     try:
-        comment_ref = db.collection_group('comments').document(comment_id)
+        posts_ref = db.collection('forum').stream()
+        comment_ref = None
+        for post in posts_ref:
+            comment_doc_ref = db.collection('forum').document(post.id).collection('comments').document(comment_id)
+            comment_doc = comment_doc_ref.get()
+            if comment_doc.exists:
+                comment_ref = comment_doc_ref
+                break
+
+        if not comment_ref:
+            raise Exception("Comment not found")
+
+        comment_data = comment_doc.to_dict()
+        if comment_data["user_id"] != user_id:
+            return jsonify({"message": "You are not authorized to delete this comment"}), 403
+
         comment_ref.delete()
         return jsonify({"message": "comment has been deleted"}), 200
     except Exception as e:
